Imprived @-tagging heuristics
This is in response to
https://github.com/TiddlySpace/tiddlyspace/issues/934

New behaviour is: Any @-tags from the post being replied to are
included, any on the tiddler (if it already exists) are included, the
space itself is included, the author of the tiddler is included, the
creator of the tiddler is included and any duplicates are removed.

That should cover everything.